Commit the changes for displaying the update branch information
a) Added new fields (i) branchName - name of the new branch created
from master and (ii) needs update - a check to confirm if this new
branch is out of date from the master branch.
(b) Used the GITHUB API to figure if the new branch is up to date with
the master. At present, deferred support has not been added for getting
dynamic information.
(c) Changed the view element for displaying the new fields.